requirements.txt
# Core
numpy>=1.24.0
ccxt>=4.0.0

# ML
scikit-learn>=1.3.0
joblib>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Telegram
python-telegram-bot>=20.0

# Utils
aiohttp>=3.9.0
python-dateutil>=2.8.0